ExchangeValidator
